Deduplicate parsing and fix parsing in emails
