+++ controller/feed.php
@@ -14,6 +14,7 @@ namespace OCA\Activity\Controller;
 
 use OCA\Activity\Data;
 use OCA\Activity\GroupHelper;
+use OCA\Activity\PlainTextParser;
 use OCA\Activity\UserSettings;
 use OCP\Activity\IManager;
 use OCP\AppFramework\Controller;
@@ -107,6 +108,7 @@ class Feed extends Controller {
 
 			// Overwrite user and language in the helper
 			$this->l = $this->l10nFactory->get('activity', $userLang);
+			$parser = new PlainTextParser($this->l);
 			$this->helper->setL10n($this->l);
 			$this->helper->setUser($user);
 
@@ -116,8 +118,8 @@ class Feed extends Controller {
 
 			$activities = [];
 			foreach ($data as $activity) {
-				$activity['subject_prepared'] = $this->parseMessage($activity['subject_prepared']);
-				$activity['message_prepared'] = $this->parseMessage($activity['message_prepared']);
+				$activity['subject_prepared'] = $parser->parseMessage($activity['subject_prepared']);
+				$activity['message_prepared'] = $parser->parseMessage($activity['message_prepared']);
 				$activities[] = $activity;
 			}
 
@@ -150,114 +152,5 @@ class Feed extends Controller {
 		}
 
 		return $response;
-	}
-
-	/**
-	 * Parse the parameters in the subject and message
-	 *
-	 * @param string $message
-	 * @return string
-	 */
-	protected function parseMessage($message) {
-		$message = $this->parseCollections($message);
-		$message = $this->parseParameters($message);
-		return $message;
-	}
-
-	/**
-	 * Parse collections
-	 *
-	 * @param string $message
-	 * @return string
-	 */
-	protected function parseCollections($message) {
-		return preg_replace_callback('/<collection>(.*?)<\/collection>/', function($match) {
-			$parameterList = explode('><', $match[1]);
-			$parameterListLength = sizeof($parameterList);
-
-			$parameters = [];
-			for ($i = 0; $i < $parameterListLength; $i++) {
-				$parameter = $parameterList[$i];
-				if ($i > 0) {
-					$parameter = '<' . $parameter;
-				}
-				if ($i + 1 < $parameterListLength) {
-					$parameter = $parameter . '>';
-				}
-
-				$parameters[] = $this->parseParameters($parameter);
-			}
-			if ($parameterListLength === 1) {
-				return array_pop($parameters);
-			} else {
-				$lastParameter = array_pop($parameters);
-				return $this->l->t('%s and %s', [
-					implode($this->l->t(', '), $parameters),
-					$lastParameter,
-				]);
-			}
-		}, $message);
-	}
-
-	/**
-	 * Parse the parameters in the subject and message
-	 *
-	 * @param string $message
-	 * @return string
-	 */
-	protected function parseParameters($message) {
-		$message = $this->parseUntypedParameters($message);
-		$message = $this->parseUserParameters($message);
-		$message = $this->parseFederatedCloudIDParameters($message);
-		$message = $this->parseFileParameters($message);
-		return $message;
-	}
-
-	/**
-	 * Display the parameter value
-	 *
-	 * @param string $message
-	 * @return string
-	 */
-	protected function parseUntypedParameters($message) {
-		return preg_replace_callback('/<parameter>(.*?)<\/parameter>/', function($match) {
-			return $match[1];
-		}, $message);
-	}
-
-	/**
-	 * Display the users display name
-	 *
-	 * @param string $message
-	 * @return string
-	 */
-	protected function parseUserParameters($message) {
-		return preg_replace_callback('/<user\ display\-name=\"(.*?)\">(.*?)<\/user>/', function($match) {
-			return $match[1];
-		}, $message);
-	}
-
-	/**
-	 * Display the full cloud id
-	 *
-	 * @param string $message
-	 * @return string
-	 */
-	protected function parseFederatedCloudIDParameters($message) {
-		return preg_replace_callback('/<federated-cloud-id\ display\-name=\"(.*?)\"\ user=\"(.*?)\"\ server=\"(.*?)\">(.*?)<\/federated-cloud-id>/', function($match) {
-			return $match[1];
-		}, $message);
-	}
-
-	/**
-	 * Display the path for files
-	 *
-	 * @param string $message
-	 * @return string
-	 */
-	protected function parseFileParameters($message) {
-		return preg_replace_callback('/<file\ link=\"(.*?)\"\ id=\"(.*?)\">(.*?)<\/file>/', function($match) {
-			return $match[3];
-		}, $message);
 	}
 }
